+++ nextstep-di/src/main/java/nextstep/di/factory/BeanFactoryUtils.java
@@ -2,6 +2,7 @@ package nextstep.di.factory;
 
 import com.google.common.collect.Sets;
 import nextstep.annotation.Inject;
+import nextstep.di.factory.exception.DoesNotAllowMultipleInjectedConstructorException;
 
 import java.lang.reflect.Constructor;
 import java.util.Map;
