+++ nextstep-di/src/main/java/nextstep/di/factory/ClassBeanDefinitionCreator.java
@@ -1,6 +1,7 @@
 package nextstep.di.factory;
 
 import com.google.common.collect.Maps;
+import nextstep.di.factory.exception.BeanCreationFailException;
 
 import java.lang.reflect.Constructor;
 import java.util.Arrays;
