+++ nextstep-di/src/main/java/nextstep/di/factory/exception/BeanCreationFailException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class BeanCreationFailException extends RuntimeException {
     private static final String MESSAGE = "빈 생성에 실패했습니다.";
