+++ nextstep-di/src/main/java/nextstep/di/factory/exception/DoesNotAllowMultipleInjectedConstructorException.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.exception;
 
 public class DoesNotAllowMultipleInjectedConstructorException extends RuntimeException {
     private static final String MESSAGE = "Inject 애너테이션은 하나의 생성자에만 선언할 수 있습니다.";
